timezone change int do decimal fix set_time_by_user_timezone function

+++ database/migrations/2021_01_03_164426_add_timezone_to_user_settings_table.php
@@ -14,7 +14,7 @@ class AddTimezoneToUserSettingsTable extends Migration
     public function up()
     {
         Schema::table('user_settings', function (Blueprint $table) {
-            $table->integer('timezone')->after('billing_phone_number')->nullable();
+            $table->decimal('timezone', 10, 1)->after('billing_phone_number')->nullable();
         });
     }
